packages/ocrbridge-easyocr/pyproject.toml
[project]
name = "ocrbridge-easyocr"
version = "3.1.1"
description = "EasyOCR engine for OCR Bridge"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "ocrbridge-core>=3.1.0",
    "easyocr>=1.7.2",
    "torch>=2.1.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
build = ["hatchling"]
test = [
    "pytest~=9.0",
    "pytest-mock>=3.12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "python-semantic-release>=9.0",
]

# Entry point for engine discovery
[project.entry-points."ocrbridge.engines"]
easyocr = "ocrbridge.engines.easyocr:EasyOCREngine"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocrbridge"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.10"
include = ["src"]
exclude = ["tests"]
extraPaths = ["typings"]
reportMissingTypeStubs = "none"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests requiring Tesseract binary",
    "slow: marks tests as slow running",
]
filterwarnings = [
    "ignore:The argument 'device' of Tensor.pin_memory.*:DeprecationWarning",
    "ignore:The argument 'device' of Tensor.is_pinned.*:DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*", "*/.venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.semantic_release]
build_command = """
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""
version_toml = ["pyproject.toml:project.version"]
